+++ src/cloudly/gcp/workflows.py
@@ -5,6 +5,7 @@ __all__ = ['Workflow', 'Execution', 'Step', 'WaitStep', 'BatchStep']
 import json
 import uuid
 from collections.abc import Sequence
+from typing import Literal
 
 from google.cloud import workflows_v1
 from google.cloud.workflows import executions_v1
@@ -169,10 +170,7 @@ class Execution:
         self._refresh()
         return str(self._execution.result)
 
-    def state(self) -> str:
-        """
-        Returned value is like "ACTIVE", "FAILED", CANCELLED", etc.
-        """
+    def state(self) -> Literal['STATE_UNSPECIFIED', 'ACTIVE', 'SUCCEEDED', 'FAILED', 'CANCELLED', 'UNAVAILABLE', 'QUEUED']:
         self._refresh()
         return self._execution.state.name
 
@@ -254,7 +252,7 @@ class Workflow:
         req = workflows_v1.GetWorkflowRequest(name=self._name)
         self._workflow = self._workflow_client().get_workflow(req)
 
-    def state(self):
+    def state(self) ->Literal['STATE_UNSPECIFIED', 'ACTIVE', 'UNAVAILABLE']:
         self._refresh()
         return self._workflow.state.name